fixed bug in session saving changes
