+++ .gitignore
@@ -14,3 +14,4 @@ __pycache__/
 
 # Logs
 *.log
+sandbox/*
