+++ app/home.py
@@ -1,17 +1,12 @@
 import os
 from dotenv import load_dotenv
+from utils import get_debater_profile
 from components.simulator import debate_simulator_page
 from components.agents_profiles import agent_profiles_page
 from components.settings import settings_page
 import streamlit as st
-import random
 
-from llm_mediator_simulation.personalities.demographics import DemographicCharacteristic
-from llm_mediator_simulation.personalities.personality import Personality
-from llm_mediator_simulation.personalities.scales import Likert3Level, Likert7AgreementLevel
-from llm_mediator_simulation.personalities.traits import PersonalityTrait
 from llm_mediator_simulation.simulation.debate.handler import DebateHandler
-from llm_mediator_simulation.simulation.debater.config import DebaterConfig, TopicOpinion
 
 from openai_key_manager import save_api_key
 
@@ -42,45 +37,10 @@ def init_session_state_vars():
         st.session_state.activate_mediator = False
 
     if "debater_model" not in st.session_state:
-        st.session_state.debater_model = "deepseek-r1:8b"
+        st.session_state.debater_model = "mistral"
 
-def get_debater_profile(agent_num):
-    name = UNISEXNAMES[agent_num]
-    # avatar = st.sidebar.text_input(f"Agent {agent_num} Avatar URL", value=f"https://via.placeholder.com/50?text=A{agent_num}")
-    # return {"name": name} # , "avatar": avatar}
-
-    personality = Personality(
-        demographic_profile={DemographicCharacteristic.ETHNICITY: random.choice(["White", "Black", "Asian", "Hispanic"]),
-                             DemographicCharacteristic.BIOLOGICAL_SEX: random.choice(["Male", "Female"]),
-                             DemographicCharacteristic.NATIONALITY: random.choice(["American"]),
-                             DemographicCharacteristic.AGE: str(random.randint(18, 80)),
-                             DemographicCharacteristic.MARITAL_STATUS: random.choice(["Single", "Married", "Divorced", "Widowed"]),
-                             DemographicCharacteristic.EDUCATION: random.choice(["High School", "College", "Graduate School"]),
-                             DemographicCharacteristic.OCCUPATION: random.choice(["Engineer", "Doctor", "Teacher", "Artist", "Unemployed", "Student"]),
-                             DemographicCharacteristic.POLITICAL_LEANING: random.choice(["Democrat", "Republican", "Independent"]),
-                             DemographicCharacteristic.RELIGION: random.choice(["Christian", "Muslim", "Jewish", "Atheist"]),
-                             DemographicCharacteristic.SEXUAL_ORIENTATION: random.choice(["Heterosexual", "Homosexual", "Bisexual"]),
-                             DemographicCharacteristic.HEALTH_CONDITION: random.choice(["Disabled", "Non-disabled"]),
-                             DemographicCharacteristic.INCOME: str(random.randint(30000, 100000)),
-                             DemographicCharacteristic.HOUSEHOLD_SIZE: str(random.randint(1, 5)),
-                             DemographicCharacteristic.NUMBER_OF_DEPENDENT: str(random.randint(0, 3)),
-                             DemographicCharacteristic.LIVING_QUARTERS: random.choice(["House", "Apartment", "Condo"]),
-                             DemographicCharacteristic.CITY_OF_RESIDENCE: random.choice(["New York", "Los Angeles", "Chicago", "Houston", "Phoenix", "Philadelphia", "San Antonio", "San Diego", "Dallas", "San Jose"]),
-                             DemographicCharacteristic.PRIMARY_MODE_OF_TRANSPORTATION: random.choice(["Car", "Public Transit", "Bicycle", "Walking"]),
-                             }
-    )
-
-    traits = {}
-    random_traits = random.sample(list(PersonalityTrait), random.randint(1, 5))
-    for trait in random_traits:
-        traits[trait] = random.choice([Likert3Level.LOW, Likert3Level.AVERAGE, Likert3Level.HIGH])
-
-    personality.traits = traits
-    return DebaterConfig(name=name,
-                         topic_opinion=TopicOpinion(agreement=random.choice(list(Likert7AgreementLevel))),
-                         personality = personality)
-    
-UNISEXNAMES = ["Alex", "Riley", "Jordan", "Parker", "Sawyer", "Taylor", "Casey", "Avery", "Jamie", "Quinn"]
+    if "remaining_rounds" not in st.session_state:
+        st.session_state.remaining_rounds = 0
 
 init_session_state_vars()
 
@@ -94,11 +54,11 @@ st.set_page_config(page_title="Debate Simulator", layout="wide")
 # Page navigation
 # st.sidebar.header("Navigation")
 page = st.sidebar.radio("Go to",
-                        ["Debate Simulator", "Agent Profiles", "Settings"])
+                        ["Debate Simulator", "Debate Settings", "OpenAI API Settings"])
 
-if page == "Agent Profiles":
+if page == "Debate Settings":
     agent_profiles_page()
 elif page == "Debate Simulator":
     debate_simulator_page()
-elif page == "Settings":
+elif page == "OpenAI API Settings":
     settings_page()
